enhancement script, code formatting
