+++ recipes/Voicebank/enhance/SGMSE/requirements_sgmse.txt
@@ -1,0 +1,25 @@
+gdown
+h5py 
+ipympl 
+librosa 
+ninja 
+numpy<2.0
+pandas 
+pesq 
+pillow 
+protobuf 
+pyarrow
+pyroomacoustics 
+pystoi 
+scipy 
+sdeint 
+setuptools 
+seaborn 
+torch 
+torch-ema 
+torch-pesq
+torchaudio
+torchvision 
+torchinfo 
+torchsde 
+tqdm
